Identifier added to role, admin role seeder

+++ app/Http/Controllers/Web/UserController.php
@@ -46,11 +46,14 @@ class UserController extends Controller
      */
     public function create()
     {
-        $roles = Role::pluck('name', 'id')->all();
-        $adminUser = new Users(auth()->user());
-        $adminUser = $adminUser->resolve();
+        $user = auth()->user();
+        if (!$user->userHasRoleByIdentifier(User::SUPER_ADMIN)) {
+            abort(403);
+        }
 
-        return view('admin.users.create', compact('roles', 'adminUser'));
+        $roles = Role::pluck('name', 'id')->all();
+
+        return view('admin.users.create', compact('roles', 'user'));
     }
 
 
@@ -62,6 +65,10 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        if (!auth()->user()->userHasRoleByIdentifier(User::SUPER_ADMIN)) {
+            abort(403);
+        }
+
         $request->validate([
             'first_name' => ['required', 'string', 'max:255'],
             'last_name' => ['required', 'string', 'max:255'],
@@ -105,6 +112,10 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        if (!auth()->user()->userHasRoleByIdentifier(User::SUPER_ADMIN)) {
+            abort(403);
+        }
+
         $user = User::find($id);
         $roles = Role::pluck('name', 'id')->all();
         $userRoles = $user->roles->pluck('name','name')->all();
@@ -121,6 +132,10 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!auth()->user()->userHasRoleByIdentifier(User::SUPER_ADMIN)) {
+            abort(403);
+        }
+
         $user = User::find($id);
 
         $request->validate([
@@ -147,6 +162,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (!auth()->user()->userHasRoleByIdentifier(User::SUPER_ADMIN)) {
+            abort(403);
+        }
+
         User::find($id)->delete();
         return redirect()->route('users.index')->with('success','Deleted successfully');
     }
